Move AdditionalPayOnyxData into IOU types and drop PayMoneyRequest re-exports
