+++ src/libs/actions/IOU/mergeAdditionalPayOnyxData.ts
@@ -1,14 +1,4 @@
-import type ONYXKEYS from '@src/ONYXKEYS';
-
-import type {OnyxUpdate} from 'react-native-onyx';
-
-type SearchPayOnyxKey = typeof ONYXKEYS.COLLECTION.RAM_ONLY_REPORT_LOADING_STATE | typeof ONYXKEYS.COLLECTION.SNAPSHOT | typeof ONYXKEYS.COLLECTION.REPORT;
-
-type AdditionalPayOnyxData = {
-    optimisticData?: Array<OnyxUpdate<SearchPayOnyxKey>>;
-    successData?: Array<OnyxUpdate<SearchPayOnyxKey>>;
-    failureData?: Array<OnyxUpdate<SearchPayOnyxKey>>;
-};
+import type AdditionalPayOnyxData from './types/AdditionalPayOnyxData';
 
 function mergeAdditionalPayOnyxData<
     T extends {
@@ -30,4 +20,3 @@ function mergeAdditionalPayOnyxData<
 }
 
 export default mergeAdditionalPayOnyxData;
-export type {AdditionalPayOnyxData, SearchPayOnyxKey};
